# api-docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MediAsk API",
    "version": "1.0.0",
    "description": "智能医疗辅助问诊系统 - 按模块组织"
  },
  "servers": [
    {
      "url": "http://localhost:8989",
      "description": "开发环境"
    }
  ],
...
  "paths": {
...
    "/api/v1/schedules": {
      "post": {
        "tags": [
          "schedule"
        ],
        "summary": "创建排班",
        "description": "手动创建单个排班；同医生+同日期+同时段已存在时返回业务错误(code=1006)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScheduleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapper"
                }
              }
            }
          }
        },
        "operationId": "postApiV1Schedules"
      },
      "get": {
        "tags": [
          "schedule"
        ],
        "summary": "分页查询排班列表",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "doctorId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "医生ID"
          },
          {
            "name": "departmentId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "科室ID（当前版本暂未生效，仅预留参数）"
          },
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "开始日期"
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "结束日期"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "enum": [
                0,
                1,
                2,
                3
              ]
            },
            "description": "状态：0=停诊，1=开放，2=约满，3=过期"
          },
          {
            "name": "pageNum",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "description": "页码",
            "example": 1
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "description": "每页数量",
            "example": 10
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResultResponse"
                }
              }
            }
          }
        },
        "operationId": "getApiV1Schedules"
      }
    },
    "/api/v1/schedules/auto": {
      "post": {
        "tags": [
          "schedule"
        ],
        "summary": "自动排班",
        "description": "仅生成并保存DRAFT方案，不直接生效；需调用发布接口使方案生效",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutoScheduleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutoSchedulePlanResponse"
                }
              }
            }
          }
        },
        "operationId": "postApiV1SchedulesAuto"
      }
    },
...
    "/api/v1/schedules/available": {
      "get": {
        "tags": [
          "schedule"
        ],
        "summary": "可预约排班",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "periodCode",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "description": "1=上午, 2=下午, 3=晚上"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapper"
                }
              }
            }
          }
        },
        "operationId": "getApiV1SchedulesAvailable"
      }
    },
    "/api/v1/schedules/batch": {
      "delete": {
        "tags": [
          "schedule"
        ],
        "summary": "批量删除排班",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "doctorId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "医生ID"
          },
          {
            "name": "startDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "开始日期"
          },
          {
            "name": "endDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "结束日期"
          },
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            },
            "description": "是否强制删除",
            "example": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchDeleteResponse"
                }
              }
            }
          }
        },
        "operationId": "deleteApiV1SchedulesBatch"
      }
    },
    "/api/v1/schedules/doctor/{doctorId}": {
      "get": {
        "tags": [
          "schedule"
        ],
        "summary": "医生排班列表",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "doctorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapper"
                }
              }
            }
          }
        },
        "operationId": "getApiV1SchedulesDoctorByDoctorId"
      }
    },
    "/api/v1/schedules/generate": {
      "post": {
        "tags": [
          "schedule"
        ],
        "summary": "根据模板生成排班",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateScheduleFromTemplateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapper"
                }
              }
            }
          }
        },
        "operationId": "postApiV1SchedulesGenerate"
      }
    },
    "/api/v1/schedules/{scheduleId}": {
      "get": {
        "tags": [
          "schedule"
        ],
        "summary": "查询排班详情",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "scheduleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapper"
                }
              }
            }
          }
        },
        "operationId": "getApiV1SchedulesByScheduleId"
      },
      "delete": {
        "tags": [
          "schedule"
        ],
        "summary": "删除排班",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "scheduleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            },
            "description": "是否强制删除（取消关联预约）",
            "example": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapper"
                }
              }
            }
          }
        },
        "operationId": "deleteApiV1SchedulesByScheduleId"
      }
    },
    "/api/v1/schedules/{scheduleId}/close": {
      "post": {
        "tags": [
          "schedule"
        ],
        "summary": "停诊",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "scheduleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "reason",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "停诊原因"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapper"
                }
              }
            }
          }
        },
        "operationId": "postApiV1SchedulesByScheduleIdClose"
      }
    },
    "/api/v1/schedules/{scheduleId}/open": {
      "post": {
        "tags": [
          "schedule"
        ],
        "summary": "开诊",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "scheduleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapper"
                }
              }
            }
          }
        },
        "operationId": "postApiV1SchedulesByScheduleIdOpen"
      }
    },
    "/api/v1/schedules/{scheduleId}/slots": {
      "put": {
        "tags": [
          "schedule"
        ],
        "summary": "调整号源",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "scheduleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "totalSlots",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapper"
                }
              }
            }
          }
        },
        "operationId": "putApiV1SchedulesByScheduleIdSlots"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "AutoScheduleRequest": {
        "type": "object",
        "required": [
          "departmentId",
          "periods"
        ],
        "properties": {
          "departmentId": {
            "type": "integer",
            "format": "int64",
            "description": "科室ID"
          },
          "dateRange": {
            "$ref": "#/components/schemas/AutoScheduleDateRangeRequest"
          },
          "doctorIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "description": "指定参与排班的医生ID列表"
          },
          "periods": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "description": "时段编码列表，1=上午 2=下午 3=晚上"
          },
          "demand": {
            "$ref": "#/components/schemas/AutoScheduleDemandRequest"
          },
          "hardConstraints": {
            "$ref": "#/components/schemas/AutoScheduleHardConstraintsRequest"
          },
          "softGoals": {
            "$ref": "#/components/schemas/AutoScheduleSoftGoalsRequest"
          },
          "solverConfig": {
            "$ref": "#/components/schemas/AutoScheduleSolverConfigRequest"
          },
          "constraintDslJson": {
            "type": "string",
            "description": "约束 DSL(JSON 字符串)，传入后优先使用 DSL 编译结果"
          },
          "ruleProfileCode": {
            "type": "string",
            "description": "规则配置编码（发布态），当 constraintDslJson 为空时自动加载对应 DSL"
          },
          "basePlanId": {
            "type": "integer",
            "format": "int64",
            "description": "增量重排基线方案ID"
          },
          "replanWindow": {
            "$ref": "#/components/schemas/AutoScheduleReplanWindowRequest"
          }
        },
        "example": {
          "departmentId": 100000000001,
          "dateRange": {
            "startDate": "2026-02-20",
            "endDate": "2026-02-26"
          },
          "doctorIds": [
            100000000001,
            100000000002
          ],
          "periods": [
            1,
            2
          ],
          "hardConstraints": {
            "maxConsecutiveDays": 5,
            "maxShiftsPerWeek": 10,
            "minRestHoursBetweenShifts": 12,
            "excludeHolidays": false,
            "holidayPolicy": "REDUCED",
            "holidayReductionFactor": 0.5
          },
          "softGoals": {
            "fairnessWeight": 0.3,
            "preferenceWeight": 0.2,
            "continuityWeight": 0.15,
            "seniorCoverageWeight": 0.2,
            "weekendBalanceWeight": 0.15
          },
          "solverConfig": {
            "strategy": "AUTO",
            "maxIterations": 2000,
            "timeLimitMs": 3000
          },
          "constraintDslJson": "{\"version\":\"1.0\",\"rules\":[],\"hardExpr\":{},\"softExpr\":{},\"objective\":{}}",
          "ruleProfileCode": "DEPT_DEFAULT",
          "basePlanId": 100000000100,
          "replanWindow": {
            "startDate": "2026-02-22",
            "endDate": "2026-02-24"
          }
        }
      },
      "AutoScheduleDateRangeRequest": {
        "type": "object",
        "required": [
          "startDate",
          "endDate"
        ],
        "properties": {
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          }
        }
      },
...
      "AutoScheduleDemandRequest": {
        "type": "object",
        "properties": {
          "byDatePeriod": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AutoScheduleDemandItemRequest"
            }
          }
        }
      },
...
      "AutoScheduleSoftGoalsRequest": {
        "type": "object",
        "properties": {
          "fairnessWeight": {
            "type": "number",
            "format": "double"
          },
          "preferenceWeight": {
            "type": "number",
            "format": "double"
          },
          "continuityWeight": {
            "type": "number",
            "format": "double"
          },
          "seniorCoverageWeight": {
            "type": "number",
            "format": "double"
          },
          "weekendBalanceWeight": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "AutoSchedulePlanResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/AutoSchedulePlanDTO"
          }
        }
      },
...
      "BatchDeleteResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "状态码"
          },
          "msg": {
            "type": "string",
            "description": "提示信息"
          },
          "data": {
            "type": "integer",
            "description": "成功删除的数量"
          }
        }
      },
...
      "CreateScheduleRequest": {
        "type": "object",
        "required": [
          "doctorId",
          "scheduleDate",
          "timePeriodCode",
          "totalSlots",
          "slotDurationMinutes"
        ],
        "properties": {
          "doctorId": {
            "type": "integer",
            "format": "int64",
            "description": "医生ID"
          },
          "scheduleDate": {
            "type": "string",
            "format": "date",
            "description": "排班日期"
          },
          "timePeriodCode": {
            "type": "integer",
            "format": "int32",
            "description": "1=上午, 2=下午, 3=晚上",
            "enum": [
              1,
              2,
              3
            ]
          },
          "totalSlots": {
            "type": "integer",
            "format": "int32",
            "description": "总号源数",
            "minimum": 1
          },
          "slotDurationMinutes": {
            "type": "integer",
            "format": "int32",
            "description": "每个号源时长（分钟）",
            "minimum": 5
          }
        },
        "example": {
          "doctorId": 1,
          "scheduleDate": "2026-02-18",
          "timePeriodCode": 1,
          "totalSlots": 20,
          "slotDurationMinutes": 15
        }
      },
...
      "GenerateScheduleFromTemplateRequest": {
        "type": "object",
        "required": [
          "templateId",
          "startDate",
          "endDate"
        ],
        "properties": {
          "templateId": {
            "type": "integer",
            "format": "int64"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          }
        }
      },
...
      "PageResultResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "状态码"
          },
          "msg": {
            "type": "string",
            "description": "提示信息"
          },
          "data": {
            "type": "object",
            "properties": {
              "total": {
                "type": "integer",
                "format": "int64",
                "description": "总记录数"
              },
              "pageNum": {
                "type": "integer",
                "format": "int32",
                "description": "页码"
              },
              "pageSize": {
                "type": "integer",
                "format": "int32",
                "description": "每页数量"
              },
              "list": {
                "type": "array",
                "description": "数据列表"
              }
            }
          }
        }
      },
...
      "ResponseWrapper": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "状态码，0表示成功"
          },
          "msg": {
            "type": "string",
            "description": "提示信息"
          },
          "data": {
            "type": "object",
            "description": "响应数据"
          },
          "traceId": {
            "type": "string",
            "description": "链路追踪ID"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64",
            "description": "响应时间戳（毫秒）"
          }
        },
        "example": {
          "code": 0,
          "msg": "success",
          "data": {}
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}